requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coverage==5.1
Django==3.0.8
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
dnspython==1.16.0
drf-schema-adapter==2.0.7
idna==2.10
importlib-metadata==1.7.0
Inflector==3.0.1
more-itertools==8.4.0
mysql-connector-python==8.0.20
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
psycopg2==2.8.5
djangorestframework-simplejwt==4.4.0
boto3==1.14.16